+++ demo/sdk_demo_v2/LOGIN_sdk_login_UI.h
@@ -136,9 +136,8 @@ public:
 
 private:
 	CVerticalLayoutUI* m_WithoutLoginRestAPIPage;
-	CRichEditUI*	   m_editRestAPIKey;
-	CRichEditUI*	   m_editRestAPISecret;
-	CRichEditUI*	   m_editEmail;
+	CRichEditUI*	   m_editRestAPIUserZAK;
+	CRichEditUI*	   m_editMeetingNumber;
 	CRichEditUI*	   m_editScreenName;
 
 	CCheckBoxUI*  m_chkRemember;
@@ -184,6 +183,7 @@ public:
 	void ChangeUIforJoinFailed();
 
 	void NotifyAuthDone();
+    void EnableEmailLoginUI(bool);
 
 protected:
 	CPaintManagerUI m_PaintManager;