+++ demo/sdk_demo_v2/SETTINGS_feature_custom_workflow.h
@@ -11,7 +11,6 @@ public:
 	void CleanUp(){}
 	void SetMeetingUIPos(ZOOM_SDK_NAMESPACE::WndPosition pos);
 	void SetFloatVideoPos(ZOOM_SDK_NAMESPACE::WndPosition pos);
-	void SetMeetingIDForMeetingUITitle(UINT64 meetingNumber);
 	void EnableLButtonDBClick4SwitchFullScreenMode(bool bEnable);
 	void EnableShareIOSDevice(bool bEnable);
 	void EnableShareWhiteBoard(bool bEnable);