+++ demo/sdk_demo_v2/schedule_meeting_ui.h
@@ -21,6 +21,7 @@ enum ErrorMessageType
 	CanNotSetPwd,
 	ApiUnknowErr,
 	LeaveEditMeeting,
+	InterpreterEmailIsEmpty,
 
 };
 enum PreMeetingPageType