+++ h/meeting_service_components/meeting_audio_interface.h
@@ -120,7 +120,7 @@ public:
 	virtual SDKError MuteAudio(unsigned int userid, bool allowUnmuteBySelf = true) = 0;
 
 	/// \brief Unmute the assigned user. 
-	/// \param userid Specify the user ID to unmute. ZERO(0) indicates to unmute all the participants. 
+	/// \param userid Specify the user ID to unmute. 
 	/// \return If the function succeeds, the return value is SDKErr_Success.
 	///Otherwise failed. To get extended error information, see \link SDKError \endlink enum.
 	/// \remarks Valid for both ZOOM style and user custom interface mode.